Camera Update and Water Obstacle
I added camera tracking to the player so now it should follow the player around and be in a much better position. This position is easily editable on the camera object.

Made the water obstacle larger and added a script that kills and respawns the player if they touch it.